+++ modules/core/src/model/input/dynamic-input.model.spec.ts
@@ -71,6 +71,12 @@ describe("DynamicInputModel test suite", () => {
             expect(defaultObject.placeholder).toBeDefined();
             expect(defaultObject.placeholder).toEqual("");
         });
+
+        it("tests if correct default readonly property is set", () => {
+
+            expect(defaultObject.readonly).toBeDefined();
+            expect(defaultObject.readonly).toBe(false);
+        });
         
         it("tests if correct default step property is set", () => {